Difficulty curve percentage is now calculated as a double rather than an integer, and the UI reflects this Upon appearing for the first time, the difficulty curve matches the sliders

+++ SetTracker/Models/DifficultyCurve.swift
@@ -19,9 +19,9 @@ class DifficultyCurve {
     }
     
     // this results in some rounding, that might mean everything does not add up to 100
-    var goalPersent: [Int : Int] {
+    var goalPersent: [Int : Double] {
         goalCount.mapValues { i in
-            Int((Double(i) / Double(total)) * 100.0)
+            (Double(i) / Double(total)) * 100.0
         }
     }
